requirements-tensorflow.txt
#
###  pip list for aruco tracking
#
# numpy (1.19.1)
# opencv-contrib-python (4.2.0.34)
# opencv-python (4.2.0.34)
# pyrealsense2 (2.35.2.1937)
# requests (2.24.0)
# gql (2.0.0)

# NOTICE for installing tensorflow
# to install tensorflow > 2.0, you need to 
# pip install --upgrade pip
# pip install setuptools --upgrade
# pip install wheel

numpy
opencv-python
opencv-contrib-python
pyrealsense2==2.35.2.1937
requests
gql
websocket-client
scipy==1.4.1
Pillow
cython
matplotlib
scikit-image==0.16.2
tensorflow==2.4.1
h5py
imgaug
IPython[all]
lxml
tqdm
absl-py
easydict
